+++ com.unity.render-pipelines.core/Runtime/RenderPipeline/IShaderVariantSettings.cs
@@ -5,11 +5,14 @@ namespace UnityEngine.Rendering
     /// </summary>
     public enum ShaderVariantLogLevel
     {
-        /// <summary>Disable all log for Shader Variant</summary>
+        // <summary>Disable all log for Shader Variant</summary>
+        [Tooltip("No shader variants are logged")]
         Disabled,
-        /// <summary>Only logs SRP Shaders when logging Shader Variant</summary>
+        /// <summary>Only logs Shaders with a Subshader containing the tag RenderPipeline</summary>
+        [Tooltip("Only shaders that are compatible with SRPs (ie: URP, HDRP) are logged")]
         OnlySRPShaders,
-        /// <summary>Logs all Shader Variant</summary>
+        /// <summary>No filter is applied to the logging of shader variants</summary>
+        [Tooltip("All shader variants are logged")]
         AllShaders,
     }